fix jointable

+++ login-service/src/main/java/com/venturedive/login/repository/UserRepository.java
@@ -1,12 +1,12 @@
 package com.venturedive.login.repository;
 
+import com.venturedive.login.entity.User;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.venturedive.login.entity.User;
-
 @Repository
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 	User findByUsernameAndPassword(String username, String password);
 }
